Added price calculation, placeholder design of price calculation

+++ Self-checkout.WpfApp/Models/ProductModel.cs
@@ -21,6 +21,6 @@ namespace Self_checkout.WpfApp.Models
             CalculatedQuantity = 1;
         }
 
-        public float CalculatedQuantity { get; set; }
+        public decimal CalculatedQuantity { get; set; }
     }
 }
